Move framet class to its own file
Having one class per file makes it easier to navigate the code and find
what you want.
